Project-root/time_management_api/requirements.txt
Flask
psycopg2-binary
azure-storage-blob